# lib/api-zod/src/generated/types/apexScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface ApexScore {
  id: number;
  runnerId: number;
  racecardId: number;
  /** 0-100 Ability rating */
  abilityScore: number;
  /** 0-100 Pace fit */
  paceFitScore: number;
  /** 0-100 Tactical resilience */
  tacticalResilienceScore: number;
  /** 0-100 Ground/trip suitability */
  groundTripScore: number;
  /** 0-100 Replay intelligence */
  replayIntelligenceScore: number;
  /** 0-100 Hidden value */
  hiddenValueScore: number;
  /** 0-100 Volatility risk (lower is better) */
  volatilityRisk: number;
  /** Computed composite APEX score */
  totalScore: number;
  /** best_of_day | top_rated_high_variance | hidden_value | replay_upgrade | no_bet */
  confidenceClass: string;
  /** @nullable */
  analystNotes?: string | null;
  /** @nullable */
  replayContext?: string | null;
  /** @nullable */
  contextualIntelligence?: string | null;
  createdAt?: string;
  updatedAt?: string;
}