# lib/api-zod/src/generated/types/calibrationEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface CalibrationEntry {
  id: number;
  date: string;
  racecardId: number;
  /** @nullable */
  runnerId?: number | null;
  predictedClass: string;
  /** win | place | unplaced | non_runner */
  outcome: string;
  /** @nullable */
  apexScore?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}