# lib/api-zod/src/generated/types/calibrationSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface CalibrationSummary {
  totalPredictions: number;
  wins: number;
  places: number;
  unplaced: number;
  strikeRate: number;
  placeRate: number;
  bestOfDayRecord?: string;
  hiddenValueRecord?: string;
}