# lib/api-zod/src/generated/types/horse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface Horse {
  id: number;
  name: string;
  /** @nullable */
  trainer?: string | null;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  sex?: string | null;
  /** @nullable */
  colour?: string | null;
  /** @nullable */
  sire?: string | null;
  /** @nullable */
  dam?: string | null;
  /** @nullable */
  preferredGoing?: string | null;
  /** @nullable */
  preferredDistance?: string | null;
  /** @nullable */
  trackSpecialistNotes?: string | null;
  /** @nullable */
  behaviourProfile?: string | null;
  /** @nullable */
  memoryNotes?: string | null;
  /** @nullable */
  replayNotes?: string | null;
  /** @nullable */
  hiddenValueFlags?: string | null;
  /** @nullable */
  volatilityRating?: string | null;
  /** @nullable */
  totalRuns?: number | null;
  /** @nullable */
  wins?: number | null;
  /** @nullable */
  places?: number | null;
  createdAt?: string;
  updatedAt?: string;
}