# lib/api-zod/src/generated/types/horseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface HorseInput {
  name: string;
  trainer?: string;
  owner?: string;
  /** @nullable */
  age?: number | null;
  sex?: string;
  colour?: string;
  sire?: string;
  dam?: string;
  preferredGoing?: string;
  preferredDistance?: string;
  trackSpecialistNotes?: string;
  behaviourProfile?: string;
  memoryNotes?: string;
  replayNotes?: string;
  hiddenValueFlags?: string;
  volatilityRating?: string;
  /** @nullable */
  totalRuns?: number | null;
  /** @nullable */
  wins?: number | null;
  /** @nullable */
  places?: number | null;
}