# lib/api-zod/src/generated/types/horseNote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface HorseNote {
  id: number;
  horseId: number;
  /** Type: memory | replay | track | context | trainer | market */
  noteType: string;
  content: string;
  /** @nullable */
  raceRef?: string | null;
  /** @nullable */
  venue?: string | null;
  /** @nullable */
  date?: string | null;
  createdAt?: string;
  updatedAt?: string;
}