# lib/api-zod/src/generated/types/horseNoteUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * APEX Racing Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface HorseNoteUpdate {
  noteType?: string;
  content?: string;
  raceRef?: string;
  venue?: string;
  date?: string;
}